add : Add login Logout Option

+++ Rooftop.WebApp/Controllers/HouseOwnerController.cs
@@ -1,9 +1,6 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-
-using Rooftop.WebApp.DatabaseContext;
 using Rooftop.WebApp.RepositoryService;
 using Rooftop.WebApp.ViewModel;
-using static System.Runtime.InteropServices.JavaScript.JSType;
 
 namespace Rooftop.WebApp.Controllers;
 
@@ -11,7 +8,7 @@ public class HouseOwnerController : Controller
 {
     private readonly IHouseOwnerRepository houseOwnerRepository;
 
-  
+
 
     public HouseOwnerController(IHouseOwnerRepository houseOwnerRepository)
     {
@@ -20,56 +17,95 @@ public class HouseOwnerController : Controller
     }
     public async Task<IActionResult> Index()
     {
-        var data = await houseOwnerRepository.GetAllAsync();
-        
-        return View(data);
-    }
-    public async Task<ActionResult<HouseOwnerVm>> CreateOrEditRoofTop(int id, CancellationToken cancellationToken)
-    {
-
-        if (id == 0)
+        var user = HttpContext.Session.GetString("adminEmail");
+        if (user != null)
         {
-            return View(new HouseOwnerVm());
-        }
-        else
-        {
-            var data = await houseOwnerRepository.GetByIdAsync(id, cancellationToken);
-            data.Password = null;
+            var data = await houseOwnerRepository.GetAllAsync();
             return View(data);
         }
 
+
+        return RedirectToAction("Login", "Admin");
+    }
+
+    public async Task<ActionResult<HouseOwnerVm>> CreateOrEditHouseOwner(int id, CancellationToken cancellationToken)
+    {
+
+        
+
+            if (id == 0)
+            {
+                return View(new HouseOwnerVm());
+            }
+            else
+            {
+                var data = await houseOwnerRepository.GetByIdAsync(id, cancellationToken);
+                data.Password = null;
+                return View(data);
+            }
+        
+
+
     }
     [HttpPost]
-    public async Task<ActionResult<HouseOwnerVm>> CreateOrEditRoofTop(int id, HouseOwnerVm houseOwnerVm, CancellationToken cancellation)
+    public async Task<ActionResult<HouseOwnerVm>> CreateOrEditHouseOwner(int id, HouseOwnerVm houseOwnerVm, CancellationToken cancellation)
     {
-        if (id == 0)
-        {
-            await houseOwnerRepository.InsertAsync(houseOwnerVm, cancellation);
-            return RedirectToAction("Index");
 
-        }
-        else
-        {
-            await houseOwnerRepository.UpdateAsync(id, houseOwnerVm, cancellation);
-            return RedirectToAction("Index");
-        }
+        
+            if (id == 0)
+            {
+                await houseOwnerRepository.InsertAsync(houseOwnerVm, cancellation);
+                return RedirectToAction("Index");
+
+            }
+            else
+            {
+                await houseOwnerRepository.UpdateAsync(id, houseOwnerVm, cancellation);
+                return RedirectToAction("Index");
+            }
+
+       
+
+
+       
+
     }
     public async Task<ActionResult<HouseOwnerVm>> Delete(int id, CancellationToken cancellation)
     {
 
-        if (id != 0)
+        var user = HttpContext.Session.GetString("adminEmail");
+        if (user != null)
         {
-            await houseOwnerRepository.DeleteAsync(id, cancellation);
-            return RedirectToAction(nameof(Index));
+
+            if (id != 0)
+            {
+                await houseOwnerRepository.DeleteAsync(id, cancellation);
+                return RedirectToAction(nameof(Index));
+            }
+            else
+            {
+                return RedirectToAction(nameof(Index));
+            }
         }
-        else
-        {
-            return RedirectToAction(nameof(Index));
-        }
+
+
+        return RedirectToAction("Login", "Admin");
+
+
     }
     public async Task<ActionResult<HouseOwnerVm>> Details(int id, CancellationToken cancellation)
-    {      
-        return View(await houseOwnerRepository.GetByIdAsync(id,cancellation));
+    {
+
+        var user = HttpContext.Session.GetString("adminEmail");
+        if (user != null)
+        {
+            return View(await houseOwnerRepository.GetByIdAsync(id, cancellation));
+
+        }
+
+
+        return RedirectToAction("Login", "Admin");
+
     }
-    
+
 }
